route된 {page}.html마다 각각의 js가 적용되도록 fix
