+++ index.html
@@ -4,6 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>POPMART</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.css" />
     <link rel="stylesheet" href="css/output/styles.css" />
   </head>
   <body>
@@ -71,5 +72,7 @@
       window.addEventListener('hashchange', loadPage);
       window.addEventListener('load', loadPage);
     </script>
+    <script src="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.js"></script>
+    <script type="module" src="js/scriptLoader.js"></script>
   </body>
 </html>
