--- js/productPageItems.js
@@ -1,161 +1,0 @@
-const items = [
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail4.png?raw=true',
-        title: '스컬판다 소리 시리즈',
-        price: '15,000원',
-    },
-
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail2.png?raw=true',
-        title: '라부부 코카콜라 시리즈',
-        price: '15,000',
-    },
-    {
-        imgSrc: 'https://github.com/oloklkl/dp-static/blob/main/popmart/thumbnail/thumbnail25.png?raw=true',
-        title: '디무 동물의 왕국 시리즈',
-        price: '15,000원',
-    },
-
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail5.png?raw=true',
-        title: '스컬판다 원더랜드 시리즈',
-        price: '15,000원',
-    },
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail6.png?raw=true',
-        title: '히로노 X 어린왕자 시리즈',
-        price: '15,000원',
-    },
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail7.png?raw=true',
-        title: '디무 꿈나라 시리즈',
-        price: '15,000원',
-    },
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail8.png?raw=true',
-        title: '피노젤리 탄생석 시리즈',
-        price: '15,000원',
-    },
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail9.png?raw=true',
-        title: '[온라인 단독] 아주라 백야의 은하계 블리스터',
-        price: '40,000원',
-    },
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail11.png?raw=true',
-        title: '메가 스페이스 몰리 스티치 400%',
-        price: '268,000원',
-    },
-
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail13.png?raw=true',
-        title: '베이비 몰리 두 살과 네 살 사이 시리즈',
-        price: '15,000원',
-    },
-
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail15.png?raw=true',
-        title: '메가 알파 스컬판다 X 에곤 실레 400%',
-        price: '300,000원',
-    },
-
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail14.png?raw=true',
-        title: '몰리 빵순이 시리즈',
-        price: '15,000원',
-    },
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail17.png?raw=true',
-        title: '스컬판다 시티 오브 나이트 시리즈',
-        price: '15,000원',
-    },
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail18.png?raw=true',
-        title: '라부부 리듬 체조 블리스터',
-        price: '40,000원',
-    },
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail26.png?raw=true',
-        title: '디무 삐에로 연주단 블리스터',
-        price: '40,000원',
-    },
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail12.png?raw=true',
-        title: '메가 스페이스 몰리 릭 앤 모티 1000%',
-        price: '1,215,000원',
-    },
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail19.png?raw=true',
-        title: '라부부 x PRONOUNCE 행운의 요정 인형',
-        price: '40,000원',
-    },
-    {
-        imgSrc: 'https://github.com/oloklkl/dp-static/blob/main/popmart/thumbnail/thumbnail1.png?raw=true',
-        title: '메가 로얄 몰리 모네 수련 400%',
-        price: '300,000원',
-    },
-
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail20.png?raw=true',
-        title: '라부부 클래식 시리즈 - 봉제인형 숄더백',
-        price: '36,000원',
-    },
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail22.png?raw=true',
-        title: '크라이베이비x파워퍼프걸 시리즈',
-        price: '50,000원 쏘쏘',
-    },
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail16.png?raw=true',
-        title: 'SKULLPANDA 스컬판다 암흑 속의 별 BJD',
-        price: '208,000원',
-    },
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail24.png?raw=true',
-        title: '히로노 X 게리 베이스맨 블리스터',
-        price: '40,000원',
-    },
-
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail27.png?raw=true',
-        title: '피노젤리 베리잼 블리스터',
-        price: '40,000원',
-    },
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail3.png?raw=true',
-        title: '크라이베이비 사랑의 눈물 시리즈 인형',
-        price: '40,000',
-    },
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail28.png?raw=true',
-        title: '피노젤리 선과 악 블리스터',
-        price: '40,000원',
-    },
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail21.png?raw=true',
-        title: '크라이베이비 날 띄워줘! 블리스터',
-        price: '40,000원',
-    },
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail29.png?raw=true',
-        title: 'THE MONSTERS x 원피스 시리즈',
-        price: '15,000원',
-    },
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail30.png?raw=true',
-        title: '바비 스타일 아이콘 시리즈',
-        price: '15,000원',
-    },
-    {
-        imgSrc: 'https://github.com/oloklkl/dp-static/blob/main/popmart/thumbnail/thumbnail10.png?raw=true',
-        title: '메가 라부부 스케치 400%',
-        price: '268,000원',
-    },
-    {
-        imgSrc: 'https://github.com/hyeonky/dp-static/blob/main/popmart/thumbnail/thumbnail23.png?raw=true',
-        title: '크라이베이비 크라잉 어게인 시리즈',
-        price: '15,000원',
-    },
-];
-
-export default items;
