+++ page/productPage/productList.html
@@ -1,39 +1,39 @@
 <!DOCTYPE html>
 <html lang="ko">
-    <head>
-        <meta charset="UTF-8" />
-        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-        <title>Popmart Collection</title>
-        <base href="/" />
-        <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.css" />
-        <link rel="stylesheet" href="resources/css/output/styles.css" />
-    </head>
-    <body>
-        <header></header>
-        <div class="inner">
-            <div class="content-wrapper">
-                <div class="brand-list">
-                    <span class="labubu">라부부s</span>
-                    <span class="crybaby">크라이베이비</span>
-                    <span class="skullpanda">스컬판다</span>
-                    <span class="physique">물리</span>
-                    <span class="pinogelly">피노젤리</span>
-                    <span class="but"></span>
-                </div>
-                <div class="grid-wrapper">
-                    <div class="grid-header">
-                        <h2>POPMART</h2>
-                        <h2>COLLECTION</h2>
-                    </div>
-                    <div class="grid-container">
-                        <div class="circle" id="circle"></div>
-                    </div>
-                </div>
-            </div>
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Popmart Collection</title>
+    <base href="/" />
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.css" />
+    <link rel="stylesheet" href="/css/output/styles.css" />
+  </head>
+  <body>
+    <header></header>
+    <div class="inner">
+      <div class="content-wrapper">
+        <div class="brand-list">
+          <span class="labubu">라부부s</span>
+          <span class="crybaby">크라이베이비</span>
+          <span class="skullpanda">스컬판다</span>
+          <span class="physique">물리</span>
+          <span class="pinogelly">피노젤리</span>
+          <span class="but"></span>
         </div>
-        <footer></footer>
+        <div class="grid-wrapper">
+          <div class="grid-header">
+            <h2>POPMART</h2>
+            <h2>COLLECTION</h2>
+          </div>
+          <div class="grid-container">
+            <div class="circle" id="circle"></div>
+          </div>
+        </div>
+      </div>
+    </div>
+    <footer></footer>
 
-        <script src="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.js"></script>
-        <script type="module" src="js/productPageScript.js"></script>
-    </body>
+    <script src="https://cdn.jsdelivr.net/npm/swiper@11/swiper-bundle.min.js"></script>
+    <script type="module" src="js/productPageScript.js"></script>
+  </body>
 </html>
